+++ tool/gec/backend/c/runtime/ge_scoop.c
@@ -38,6 +38,52 @@ extern "C" {
  */
 static EIF_MUTEX_TYPE* GE_scoop_multisessions_mutex;
 
+
+/*
+ * Passive region to be handled by `GE_process_scoop_passive_regions`.
+ */
+static GE_scoop_region* GE_current_scoop_passive_region;
+
+/* 
+ * Mutex to set and access `GE_current_scoop_passive_region'.
+ */
+static EIF_MUTEX_TYPE* GE_scoop_passive_region_mutex;
+
+/* 
+ * Condition variable to wait for the next passive region to be handled.
+ */
+static EIF_COND_TYPE* GE_scoop_passive_region_condition_variable;
+
+/* 
+ * Mutex to initialize passive region processing.
+ */
+static EIF_MUTEX_TYPE* GE_scoop_init_passive_region_mutex;
+
+/* 
+ * Condition variable to initialize passive region processing.
+ */
+static EIF_COND_TYPE* GE_scoop_init_passive_region_condition_variable;
+
+/*
+ * If the processor of `a_region` was waiting for a session to execute,
+ * then wake it up.
+ * 
+ * Not thread-safe.
+ * Need to be protected by:
+ * - `a_region->mutex`.
+ */
+static void GE_unprotected_scoop_region_wakeup(GE_scoop_region* a_region)
+{
+	if (a_region->is_passive) {
+		GE_mutex_lock((EIF_POINTER)GE_scoop_passive_region_mutex);
+		GE_current_scoop_passive_region = a_region;
+		GE_condition_variable_broadcast((EIF_POINTER)GE_scoop_passive_region_condition_variable);
+		GE_mutex_unlock((EIF_POINTER)GE_scoop_passive_region_mutex);
+	} else {
+		GE_condition_variable_broadcast((EIF_POINTER)a_region->condition_variable);
+	}
+}
+
 /* 
  * Number SCOOP sessions not fully executed yet.
  *
@@ -115,7 +161,7 @@ uint32_t GE_decrement_scoop_sessions_count()
 			*/
 		l_main_region = GE_main_context->region;
 		GE_mutex_lock((EIF_POINTER)l_main_region->mutex);
-		GE_condition_variable_broadcast((EIF_POINTER)l_main_region->condition_variable);
+		GE_unprotected_scoop_region_wakeup(l_main_region);
 		GE_mutex_unlock((EIF_POINTER)l_main_region->mutex);
 	}
 	return l_result;
@@ -180,8 +226,6 @@ static void GE_scoop_region_destroy(GE_scoop_region* a_region)
 {
 	GE_mutex_destroy((EIF_POINTER)a_region->mutex);
 	GE_condition_variable_destroy((EIF_POINTER)a_region->condition_variable);
-	GE_mutex_destroy((EIF_POINTER)a_region->sync_mutex);
-	GE_condition_variable_destroy((EIF_POINTER)a_region->sync_condition_variable);
 	GE_mutex_destroy((EIF_POINTER)a_region->precondition_mutex);
 	GE_condition_variable_destroy((EIF_POINTER)a_region->precondition_condition_variable);
 	GE_free(a_region->keep_alive);
@@ -200,7 +244,7 @@ static void GE_scoop_region_dispose(void* a_region, void* data)
 	if (l_context && l_context->region == l_region) {
 		l_context->is_region_alive = 0;
 		if (GE_mutex_try_lock((EIF_POINTER)l_region->mutex)) {
-			GE_condition_variable_broadcast((EIF_POINTER)l_region->condition_variable);
+			GE_unprotected_scoop_region_wakeup(l_region);
 			GE_mutex_unlock((EIF_POINTER)l_region->mutex);
 		}
 	} else if (l_region->is_passive) {
@@ -222,8 +266,6 @@ GE_scoop_region* GE_new_scoop_region(GE_context* a_context, char a_is_passive)
 	l_region->keep_alive = (GE_scoop_region**)GE_calloc_uncollectable(1, sizeof(GE_scoop_region*));
 	l_region->mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
 	l_region->condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
-	l_region->sync_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
-	l_region->sync_condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
 	l_region->precondition_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
 	l_region->precondition_condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
 	l_region->is_impersonation_allowed = '\1';
@@ -316,9 +358,7 @@ static void GE_unprotected_add_scoop_session(GE_scoop_session* a_session)
 	}
 	l_region->last_session = a_session;
 	*(l_region->keep_alive) = l_region;
-	/* If the processor of `l_region` was waiting for a session to execute,
-	 * then wake it up. */
-	GE_condition_variable_broadcast((EIF_POINTER)l_region->condition_variable);
+	GE_unprotected_scoop_region_wakeup(l_region);
 }
 
 /* 
@@ -376,7 +416,7 @@ static void GE_add_scoop_multisession(GE_scoop_session* a_session)
  * - `a_session->callee->mutex`.
  * - `a_session->mutex` to set `a_session->is_submitted`.
  */
-void GE_add_scoop_session(GE_scoop_session* a_session)
+static void GE_add_scoop_session(GE_scoop_session* a_session)
 {
 	GE_scoop_region* l_region;
 
@@ -397,7 +437,8 @@ void GE_add_scoop_session(GE_scoop_session* a_session)
 /* 
  * Remove SCOOP session `a_session' from the list of sessions to be executed by the processor of its callee.
  *
- * To be executed by the thread associated with the callee of `a_session'.
+ * To be executed by the thread associated with the callee of `a_session' or the
+ * by the thread handling passive regions.
  * 
  * Thread-safe.
  * Protected by:
@@ -476,7 +517,8 @@ static void GE_unprotected_move_scoop_session_to_first(GE_scoop_session* a_sessi
  * of the callee regions. If so, move them to the first position in the list of
  * sessions of their respective region and mark them as ready for execution.
  * 
- * To be executed by the thread associated with `a_region'.
+ * To be executed by the thread associated with `a_region' or the
+ * by the thread handling passive regions.
  * 
  * Thread-safe.
  * Protected by:
@@ -522,7 +564,7 @@ static void GE_promote_scoop_session(GE_scoop_region* a_region)
 				}
 				if (l_sibling != l_session) {
 					l_sibling_region = l_sibling->callee;
-					GE_condition_variable_broadcast((EIF_POINTER)l_sibling_region->condition_variable);
+					GE_unprotected_scoop_region_wakeup(l_sibling_region);
 					GE_mutex_unlock((EIF_POINTER)l_sibling_region->mutex);
 				}
 				l_sibling = l_sibling->next_sibling_session;
@@ -538,23 +580,27 @@ static void GE_promote_scoop_session(GE_scoop_region* a_region)
 }
 
 /*
- * Is `a_session` the first session submitted for execution to the processor of its callee's region?
- *
- * To be executed by the thread associated with the caller of `a_session'.
- *
- * Thread-safe.
- * Protected by:
- * - `a_session->callee->mutex`.
+ * New of SCOOP synchronization.
  */
-static char GE_scoop_session_is_first(GE_scoop_session* a_session)
+GE_scoop_synchronization* GE_new_scoop_synchronization()
 {
-	char l_result;
-	GE_scoop_region* l_callee = a_session->callee;
+	GE_scoop_synchronization* l_synchronization;
 
-	GE_mutex_lock((EIF_POINTER)l_callee->mutex);
-	l_result = (l_callee->first_session == a_session);
-	GE_mutex_unlock((EIF_POINTER)l_callee->mutex);
-	return l_result;
+	l_synchronization = (GE_scoop_synchronization*)GE_calloc(1, sizeof(GE_scoop_synchronization));
+	l_synchronization->wait = '\1';
+	l_synchronization->mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
+	l_synchronization->condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
+	return l_synchronization;
+}
+
+/*
+ * Free memory allocated by `a_synchronization'.
+ */
+void GE_free_scoop_synchronization(GE_scoop_synchronization* a_synchronization)
+{
+	GE_mutex_destroy((EIF_POINTER)a_synchronization->mutex);
+	GE_condition_variable_destroy((EIF_POINTER)a_synchronization->condition_variable);
+	GE_free(a_synchronization);
 }
 
 /* 
@@ -567,6 +613,9 @@ GE_scoop_call* GE_new_scoop_call(GE_scoop_region* a_caller, char a_is_synchronou
 	l_call = (GE_scoop_call*)GE_calloc(1, a_size);
 	l_call->caller = a_caller;
 	l_call->is_synchronous = a_is_synchronous;
+	if (a_is_synchronous) {
+		l_call->synchronization = GE_new_scoop_synchronization();
+	}
 	l_call->execute = a_execute;
 	return l_call;
 }
@@ -585,6 +634,9 @@ static void GE_free_scoop_call(GE_scoop_call* a_call)
  * To be executed by the thread associated with the caller of `a_session'.
  * 
  * Thread-safe.
+ * Protected by:
+ * - `a_session->mutex`.
+ * - `a_session->callee->mutex`.
  */
 void GE_scoop_session_add_call(GE_scoop_session* a_session, GE_scoop_call* a_call)
 {
@@ -592,10 +644,8 @@ void GE_scoop_session_add_call(GE_scoop_session* a_session, GE_scoop_call* a_cal
 	GE_scoop_region* l_caller = a_call->caller;
 	GE_scoop_region* l_callee = a_session->callee;
 	char l_is_synchronous = a_call->is_synchronous;
+	GE_scoop_synchronization* l_synchronization = a_call->synchronization;
 
-	if (l_is_synchronous) {
-		GE_mutex_lock((EIF_POINTER)l_caller->sync_mutex);
-	}
 	GE_mutex_lock((EIF_POINTER)a_session->mutex);
 	l_last_call = a_session->last_call;
 	a_session->last_call = a_call;
@@ -608,19 +658,24 @@ void GE_scoop_session_add_call(GE_scoop_session* a_session, GE_scoop_call* a_cal
 		/* First call added to `a_session'. */
 		GE_mutex_unlock((EIF_POINTER)a_session->mutex);
 		GE_add_scoop_session(a_session);
+	} else if (l_callee->is_passive) {
+		GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+		GE_mutex_lock(l_callee->mutex);
+		GE_unprotected_scoop_region_wakeup(l_callee);
+		GE_mutex_unlock(l_callee->mutex);
 	} else {
 		GE_condition_variable_broadcast((EIF_POINTER)a_session->condition_variable);
 		GE_mutex_unlock((EIF_POINTER)a_session->mutex);
 	} 
 	if (l_is_synchronous) {
-		char l_wait = '\1';
-		if (l_callee->is_passive) {
-			l_wait = !GE_scoop_session_is_first(a_session);
+		if (l_synchronization) {
+			GE_mutex_lock((EIF_POINTER)l_synchronization->mutex);
+			if (l_synchronization->wait) {
+				GE_condition_variable_wait((EIF_POINTER)l_synchronization->condition_variable, (EIF_POINTER)l_synchronization->mutex);
+			}
+			GE_mutex_unlock((EIF_POINTER)l_synchronization->mutex);
+			GE_free_scoop_synchronization(l_synchronization);
 		}
-		if (l_wait) {
-			GE_condition_variable_wait((EIF_POINTER)l_caller->sync_condition_variable, (EIF_POINTER)l_caller->sync_mutex);
-		}
-		GE_mutex_unlock((EIF_POINTER)l_caller->sync_mutex);
 		if (l_callee->progagate_exception) {
 			l_callee->progagate_exception = 0;
 			/* The exception is propagated to the caller region. */
@@ -644,9 +699,58 @@ void GE_scoop_session_add_sync_call(GE_scoop_region* a_caller, GE_scoop_session*
 {
 	GE_scoop_call* l_call;
 
-	if (!GE_scoop_session_is_synchronized(a_session) && (!a_session->callee->is_passive || !GE_scoop_session_is_submitted(a_session))) {
+	if (!GE_scoop_session_is_synchronized(a_session) && (!a_session->callee->is_passive || !GE_scoop_session_is_running(a_session))) {
 		l_call = GE_new_scoop_call(a_caller, '\1', 0, sizeof(GE_scoop_call));
 		GE_scoop_session_add_call(a_session, l_call);
+	}
+}
+
+/* 
+ * Add a SCOOP call to `a_session' to indicate that `a_session` is running.
+ *
+ * To be executed by the thread associated with `a_caller'.
+ * 
+ * Thread-safe.
+ * Protected by:
+ * - `a_session->mutex`.
+ * - `a_session->callee->mutex`.
+ */
+void GE_scoop_session_add_running_call(GE_scoop_region* a_caller, GE_scoop_session* a_session, GE_scoop_synchronization* a_synchronization)
+{
+	GE_scoop_call* l_call;
+	GE_scoop_region* l_callee = a_session->callee;
+
+	GE_mutex_lock((EIF_POINTER)l_callee->mutex);
+	if (a_session->is_running) {
+		GE_mutex_unlock((EIF_POINTER)l_callee->mutex);
+		GE_mutex_lock((EIF_POINTER)a_synchronization->mutex);
+		a_synchronization->wait = 0;
+		GE_condition_variable_broadcast((EIF_POINTER)a_synchronization->condition_variable);
+		GE_mutex_unlock((EIF_POINTER)a_synchronization->mutex);
+	} else {
+		l_call = GE_new_scoop_call(a_caller, 0, 0, sizeof(GE_scoop_call));
+		l_call->is_synchronous = '\1';
+		l_call->synchronization = a_synchronization;
+		GE_mutex_lock((EIF_POINTER)a_session->mutex);
+		l_call->next = a_session->first_call;
+		a_session->first_call = l_call;
+		if (!a_session->last_call) {
+			a_session->last_call = l_call;
+		}
+		if (!a_session->is_submitted) {
+			/* First call added to `a_session'. */
+			GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+			GE_mutex_unlock((EIF_POINTER)l_callee->mutex);
+			GE_add_scoop_session(a_session);
+		} else if (l_callee->is_passive) {
+			GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+			GE_unprotected_scoop_region_wakeup(l_callee);
+			GE_mutex_unlock(l_callee->mutex);
+		} else {
+			GE_condition_variable_broadcast((EIF_POINTER)a_session->condition_variable);
+			GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+			GE_mutex_unlock((EIF_POINTER)l_callee->mutex);
+		}
 	}
 }
 
@@ -877,21 +981,20 @@ uint32_t GE_scoop_session_is_open(GE_scoop_session* a_session)
 }
 
 /*
- * Has `a_session` been submitted for execution to the processor of its callee's region?
+ * Is `a_session` already running?
  *
  * To be executed by the thread associated with the caller of `a_session'.
  * 
  * Thread-safe.
  * Protected by:
- * - `a_session->mutex`.
+ * - `a_session->callee->mutex`.
  */
-char GE_scoop_session_is_submitted(GE_scoop_session* a_session)
-{
+char GE_scoop_session_is_running(GE_scoop_session* a_session){
 	char l_result;
 
-	GE_mutex_lock((EIF_POINTER)a_session->mutex);
-	l_result = a_session->is_submitted;
-	GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+	GE_mutex_lock((EIF_POINTER)a_session->callee->mutex);
+	l_result = a_session->is_running;
+	GE_mutex_unlock((EIF_POINTER)a_session->callee->mutex);
 	return l_result;
 }
 
@@ -929,6 +1032,8 @@ static void GE_scoop_call_execute(GE_context* a_context, GE_scoop_session* a_ses
 	GE_context* l_caller_context;
 	GE_context l_old_context;
 	char l_is_synchronous = a_call->is_synchronous;
+	GE_scoop_synchronization* l_synchronization = a_call->synchronization;
+	char l_already_synchronized = 0;
 	GE_rescue r;
 
 	l_caller = a_call->caller;
@@ -945,6 +1050,7 @@ static void GE_scoop_call_execute(GE_context* a_context, GE_scoop_session* a_ses
 			a_context->last_rescue = &r;
 			if (GE_setjmp(r.jb) != 0) {
 				a_context->last_rescue = &r;
+				GE_scoop_session_set_eiffel_called(a_session, '\1');
 				if (l_is_synchronous) {
 					/* The exception will be propagated to the caller region. */
 					GE_append_to_exception_trace_buffer(&l_caller_context->last_exception_trace, a_context->last_exception_trace.area);
@@ -954,9 +1060,12 @@ static void GE_scoop_call_execute(GE_context* a_context, GE_scoop_session* a_ses
 					GE_wipe_out_exception_trace_buffer(&a_context->exception_trace_buffer);
 					GE_scoop_region_release_locks(l_caller, l_callee);
 					l_callee->progagate_exception = '\1';
-					GE_mutex_lock((EIF_POINTER)l_caller->sync_mutex);
-					GE_condition_variable_broadcast((EIF_POINTER)l_caller->sync_condition_variable);
-					GE_mutex_unlock((EIF_POINTER)l_caller->sync_mutex);
+					if (!l_already_synchronized && l_synchronization) {
+						GE_mutex_lock((EIF_POINTER)l_synchronization->mutex);
+						l_synchronization->wait = 0;
+						GE_condition_variable_broadcast((EIF_POINTER)l_synchronization->condition_variable);
+						GE_mutex_unlock((EIF_POINTER)l_synchronization->mutex);
+					}
 				} else {
 					l_callee->is_dirty = '\1';
 #ifdef GE_SCOOP_EXCEPTIONS_STOP_WHEN_DIRTY
@@ -979,10 +1088,21 @@ static void GE_scoop_call_execute(GE_context* a_context, GE_scoop_session* a_ses
 			GE_scoop_session_set_eiffel_called(a_session, '\1');
 		}
 	}
-	if (l_is_synchronous) {
-		GE_mutex_lock((EIF_POINTER)l_caller->sync_mutex);
-		GE_condition_variable_broadcast((EIF_POINTER)l_caller->sync_condition_variable);
-		GE_mutex_unlock((EIF_POINTER)l_caller->sync_mutex);
+	if (l_synchronization) {
+		if (l_is_synchronous) {
+			GE_mutex_lock((EIF_POINTER)l_synchronization->mutex);
+			l_synchronization->wait = 0;
+			GE_condition_variable_broadcast((EIF_POINTER)l_synchronization->condition_variable);
+			l_already_synchronized = '\1';
+			GE_mutex_unlock((EIF_POINTER)l_synchronization->mutex);
+		} else {
+			GE_mutex_lock((EIF_POINTER)l_synchronization->mutex);
+			if (l_synchronization->wait) {
+				GE_condition_variable_wait((EIF_POINTER)l_synchronization->condition_variable, (EIF_POINTER)l_synchronization->mutex);
+			}
+			GE_mutex_unlock((EIF_POINTER)l_synchronization->mutex);
+			GE_free_scoop_synchronization(l_synchronization);
+		}
 	}
 }
 
@@ -998,15 +1118,13 @@ static void GE_scoop_session_execute(GE_context* a_context, GE_scoop_session* a_
 		GE_mutex_lock((EIF_POINTER)a_session->mutex);
 		l_call = a_session->first_call;
 		if (l_call) {
-			GE_mutex_unlock((EIF_POINTER)a_session->mutex);
-			GE_scoop_call_execute(a_context, a_session, l_call);
-			GE_mutex_lock((EIF_POINTER)a_session->mutex);
 			l_next_call = l_call->next;
 			a_session->first_call = l_next_call;
 			if (!l_next_call) {
 				a_session->last_call = 0;
 			}
 			GE_mutex_unlock((EIF_POINTER)a_session->mutex);
+			GE_scoop_call_execute(a_context, a_session, l_call);
 			GE_free_scoop_call(l_call);
 		} else if (a_session->is_open && !a_session->callee->is_passive) {
 				/* The session is not finished. New calls may still be registered. */
@@ -1089,7 +1207,7 @@ void GE_scoop_session_close(GE_scoop_region* a_caller, GE_scoop_session* a_sessi
 	GE_scoop_session* l_locked_session;
 	GE_scoop_session* l_next_session;
 	GE_scoop_session* l_next_pending_session;
-	GE_scoop_region* l_callee;
+	GE_scoop_region* l_callee = a_session->callee;
 	char l_to_be_freed = 0;
 	char l_mutex_unlocked = 0;
 
@@ -1111,37 +1229,16 @@ void GE_scoop_session_close(GE_scoop_region* a_caller, GE_scoop_session* a_sessi
 			a_session->next_locked_session = 0;
 			if (!a_session->is_submitted) {
 				l_to_be_freed = '\1';
-			} else if (a_session->callee->is_passive) {
+			} else if (l_callee->is_passive) {
 				/* A passive region as no associated processor. Its sessions are synchronously 
 				 * executed by the processor of their callers. So the caller's processor has
 				 * to do the house cleaning (e.g. call to `GE_remove_scoop_session`).
 				 */
-				l_callee = a_session->callee;
 				GE_mutex_unlock((EIF_POINTER)a_session->mutex);
 				l_mutex_unlocked = '\1';
-				GE_remove_scoop_session(a_session);
-				/* It also has to tell any subsequent caller processor waiting for this region
-				 * to be available that we don't use it anymore. 
-				 * When the session's callee is passive, `GE_scoop_session_execute` just wakes
-				 * up its caller's processor and exits. */
 				GE_mutex_lock(l_callee->mutex);
-				l_next_session = l_callee->first_session;
-				if (l_next_session && !l_next_session->is_running && (l_next_session->next_sibling_session != l_next_session)) {
-					GE_mutex_unlock(l_callee->mutex);
-					GE_promote_scoop_session(l_callee);
-					GE_mutex_lock(l_callee->mutex);
-					l_next_session = l_callee->first_session;
-					if (!l_next_session->is_running && (l_next_session->next_sibling_session != l_next_session)) {
-						l_next_session = 0;
-					}
-				}
-				if (l_next_session) {
-					l_next_session->is_running = '\1';
-					GE_mutex_unlock(l_callee->mutex);
-					GE_scoop_session_execute(0, l_next_session);
-				} else {
-					GE_mutex_unlock(l_callee->mutex);
-				}
+				GE_unprotected_scoop_region_wakeup(l_callee);
+				GE_mutex_unlock(l_callee->mutex);
 			} else {
 				/* Wake up the callee's processor if needed to tell it that there is no call
 				 * to be added anymore. */
@@ -1326,6 +1423,68 @@ void GE_scoop_processor_run(GE_context* a_context)
 	}
 }
 
+/*
+ * Handle the sessions of passive region `a_region`.
+ * To be run in a dedicated thread. 
+ */
+static void GE_process_scoop_passive_region(GE_scoop_region* a_region)
+{
+	EIF_POINTER l_mutex = a_region->mutex;
+	GE_scoop_session* l_session;
+
+	while (1) {
+		GE_mutex_lock(l_mutex);
+		l_session = a_region->first_session;
+		if (l_session && !l_session->is_running && (l_session->next_sibling_session != l_session)) {
+			GE_mutex_unlock(l_mutex);
+			GE_promote_scoop_session(a_region);
+			GE_mutex_lock(l_mutex);
+			l_session = a_region->first_session;
+			if (l_session && !l_session->is_running && (l_session->next_sibling_session != l_session)) {
+				l_session = 0;
+			}
+		}
+		if (l_session) {
+			l_session->is_running = '\1';
+			GE_mutex_unlock(l_mutex);
+			GE_scoop_session_execute(0, l_session);
+			if (!l_session->is_open) {
+				GE_remove_scoop_session(l_session);
+			} else {
+				break;
+			}
+		} else {
+			GE_mutex_unlock(l_mutex);
+			break;
+		}
+	}
+}
+
+/*
+ * Handle the sessions of passive regions.
+ * To be run in a dedicated thread. 
+ */
+void GE_process_scoop_passive_regions()
+{
+	GE_scoop_region* l_region = 0;
+
+	GE_mutex_lock(GE_scoop_passive_region_mutex);
+		/* The thread is now ready to handle passive regions. */
+	GE_mutex_lock(GE_scoop_init_passive_region_mutex);
+	GE_condition_variable_broadcast((EIF_POINTER)GE_scoop_init_passive_region_condition_variable);
+	GE_mutex_unlock(GE_scoop_init_passive_region_mutex);
+	while (1) {
+		GE_condition_variable_wait(GE_scoop_passive_region_condition_variable, GE_scoop_passive_region_mutex);
+		l_region = GE_current_scoop_passive_region;
+		GE_current_scoop_passive_region = 0;
+		if (l_region) {
+			GE_process_scoop_passive_region(l_region);
+			l_region = 0;
+		}
+	}
+	GE_mutex_unlock(GE_scoop_passive_region_mutex);
+}
+
 /* 
  * Initialization of SCOOP.
  */
@@ -1334,6 +1493,19 @@ void GE_init_scoop()
 	GE_unprotected_scoop_sessions_count = 0;
 	GE_scoop_sessions_count_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
 	GE_scoop_multisessions_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
+	GE_current_scoop_passive_region = 0;
+	GE_scoop_passive_region_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
+	GE_scoop_passive_region_condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
+	GE_scoop_init_passive_region_mutex = (EIF_MUTEX_TYPE*)GE_mutex_create();
+	GE_scoop_init_passive_region_condition_variable = (EIF_COND_TYPE*)GE_condition_variable_create();
+	/* Make sure that we don't try tp handle passive regions before
+	 * the corresponding thread has been started and is ready. */
+	GE_mutex_lock(GE_scoop_init_passive_region_mutex);
+	GE_scoop_passive_regions_thread_create_with_attr(0);
+	GE_condition_variable_wait(GE_scoop_init_passive_region_condition_variable, GE_scoop_init_passive_region_mutex);
+	GE_mutex_unlock(GE_scoop_init_passive_region_mutex);
+	GE_mutex_destroy((EIF_POINTER)GE_scoop_init_passive_region_mutex);
+	GE_condition_variable_destroy((EIF_POINTER)GE_scoop_init_passive_region_condition_variable);
 }
 
 #ifdef __cplusplus